requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
faiss-cpu
unstructured
tiktoken
pickle-mixin
python-dotenv
